+++ xyui/avalonia/gallery/XYUI.Avalonia.Gallery/Views/XYUI1ComponentDocumentView.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace XYUI.Avalonia.Gallery.Views;
+
+public partial class XYUI1ComponentDocumentView : UserControl
+{
+    public XYUI1ComponentDocumentView() => InitializeComponent();
+}