+++ xyui/avalonia/gallery/XYUI.Avalonia.Gallery/Views/XYUI1DocumentationView.axaml.cs
@@ -1,0 +1,12 @@
+using Avalonia.Controls;
+
+namespace XYUI.Avalonia.Gallery.Views;
+
+public partial class XYUI1DocumentationView : UserControl
+{
+    public XYUI1DocumentationView()
+    {
+        InitializeComponent();
+        DataContext = new XYUI1DocumentationViewModel();
+    }
+}